LangChain/Tourist Guide/requirements.txt
langchain
google-generativeai
huggingface_hub
langchain-google-genai
langchain_community
python-dotenv
streamlit